combine uri feature with main branch
