--- src/URI/URIData.hpp
@@ -1,23 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <string>
-#include <map>
-#include <vector>
-
-class URIData{
-private:
-	std::vector<std::string> _path;
-	std::map<std::string, std::string> _query;
-
-public:
-	URIData();
-	~URIData();
-	URIData(const URIData &other);
-	const URIData &operator=(const URIData &other);
-
-	void set_path(std::vector<std::string> path);
-	void set_query(std::map<std::string, std::string> query);
-	const std::vector<std::string> get_path(void) const;
-	const std::map<std::string, std::string> get_query(void) const;
-};
