--- src/URI/Utils.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <vector>
-#include <string>
-
-class Utils{
-public:
-	static std::vector<std::string> _split_line(const std::string& line, const char delimiter);
-
-};
